New: Use Ranges and Services for item information
Previously, information on the contents of a page were given using an AnnotationList. This was problematic because technically structured data was not allowed in an OpenAnnotation annotation.

This commit changes this behaviour by introducing sequences and ranges to deliver page-level contents, and a third-party service block to carry extra information on the contents of the page.

Closes #17