New: Service blocks are no longer embedded in manifests
Previously, all information about pieces on pages were embedded in the IIIF manifest. This led to a lot of information needing to be sent over the wire on an initial download.

This commit changes this behaviour by moving the DIAMM service block to a resolveable URI. It adds a new structure serializer that contains a minimal service block which points to a resolvable representation of the information contained in the service, and this minimal serializer is used in the manifest.

Resolving the service block will do a lookup in Solr to provide the DIAMM information for that particular item (e.g., voice, composer, etc.)

Also contained in this commit is a view for resolving the service data.

Fixes #48