Fixed: IIIF Manifests now use members instead of canvases
Previously the IIIF manifest builder would produce an array of canvas URIs. The canvas key is set to be deprecated in IIIF 3.0, so we are transitioning to use the `members` key only.

This commit also introduces a change to the solr indexing serializer. In order to support the `members` functionality we need to know the page label to supply the required 'label' field. A new Solr field, `pages_ssni` (stored not indexed) stores a combination of PK and label for each page on the item record so that we save an extra lookup from Solr for the page label.

Fixes #52