Refactor: Source type filtering
This commit refactors source type filtering with images.

The behaviour of source image filtering from the user's standpoint was that clicking on a filter removed all the other options. So clicking on 'composition' removed the option for 'source'.

This commit uses Solr's multi-facet functionality to preserve the facets across filtering, so that the counts of the results, and the option to switch facets in the interface, stays the same.

This commit also refactors template usage. Previously two templates were used to render the links: One for the sources with images (image_links), and one for everything else (type_links). This commit refactors them down to a single template.

Refs #37