Fixed: Re-work contribution templates
This commit re-works and simplifies the contributions block.

The templates have been condensed into a single template and moved to a 'contribution.jinja2' file inside of 'blocks'. (The distinction between pages and blocks is that pages are extended, whereas blocks are included.)

The HTML structure for the contribution form was also standardized and included in the block. This helps style it consistently across all the different templates that have contribution functionality.

Refs #32