Fixed: Remove is_ajax request
The behaviour of this function is undefined with this check if the request is not an ajax request. Since we don't really care if it is an AJAX request, but only that it's a POST, it's not necessary to nest the rest of this function inside of that `if` statement.