Fixed: Case sensitivity in search results
A search for "Troyes" would return a different set of results than a search for "troyes." It turns out this is because the boosting fields for results were not equipped with a LowerCaseFilterFactory filter, so the results for upper-case results would be boosted over lowercase.

This commit adds LowerCaseFilterFactory filters to the text_general and text_en field types.

Fixes #103