Fix: Fixes issue with diva dimensions
In the source-detail view, Diva is being instantiated in a div that is
not yet displayed. When this div becomes visible, the diva viewer
doesn't recognize it has changed dimensions and causes issues rendering
tiles.

The temporary solution is to trigger a window resize even on tab
changes

Diva 5 should resolve this issue