Refactor the century list method
Previously the century list method had a triple nested loop. Now it
makes use of a recursive call to build the century intervals and then
sums the results.

Note: This doesn't bring down the time complexity and I'm not sure it's
any more readable. This commit may be a candidate for a revert