Fixed: Contribute link shows for every user
This commit makes it so the 'contribute' button shows up for non-logged-in users. It also fixes a bug with the images tab, where clicking on the tabbed when logged out would just bring you back to the source record, not the images tab. Now clicking on a tab, and then logging in, will redirect you back to the view you wanted in the first place.

Fixes #275